app/requirements.txt
APScheduler==3.10.4
Authlib==1.3.1
beartype==0.18.2
blinker==1.7.0
boolean.py==4.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.4
DateTime==5.4
Flask==3.0.2
Flask-JWT-Extended==4.6.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
idna==3.7
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.4
jsonpath-ng==1.6.1
license-expression==30.3.0
MarkupSafe==2.1.4
pillow==10.3.0
ply==3.11
psycopg2-binary==2.9.9
pycparser==2.21
PyJWT==2.8.0
pyotp==2.9.0
pyparsing==3.1.2
python-dotenv==1.0.1
python-magic==0.4.27
pytz==2024.1
PyYAML==6.0.1
rdflib==7.0.0
requests==2.32.2
semantic-version==2.10.0
setuptools==70.0.0
six==1.16.0
spdx-tools==0.8.2
SQLAlchemy==2.0.25
typing_extensions==4.9.0
tzlocal==5.2
uritools==4.0.2
urllib3==2.2.2
uuid==1.30
Werkzeug==3.0.3
WTForms==3.1.2
xmltodict==0.13.0
zope.interface==6.1
